api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
redis==5.2.0
rq==1.16.2
python-multipart==0.0.12
python-dateutil==2.9.0.post0
orjson==3.10.11
pytest==8.3.4
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
